@@ -1,6 +1,6 @@
 {
   "name": "make-it-beautiful",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "description": "Make text pages beautiful and easy to read",
   "repository": "https://github.com/Lifeni/make-it-beautiful",
   "author": "Lifeni",
@@ -8,23 +8,32 @@
   "private": true,
   "scripts": {
     "build:all": "npm run build && npm run build:v2",
-    "build": "webpack --config webpack.config.js --progress",
-    "build:v2": "webpack --config webpack.config.v2.js --progress",
-    "dev": "webpack --config webpack.config.js --watch --progress"
+    "build": "webpack --config webpack.config.ts --progress",
+    "build:v2": "webpack --config webpack.config.v2.ts --progress",
+    "dev": "webpack --config webpack.config.ts --watch --progress"
   },
   "devDependencies": {
+    "@types/chrome": "^0.0.130",
+    "@types/codemirror": "0.0.108",
+    "@types/copy-webpack-plugin": "^6.4.0",
+    "@types/file-saver": "^2.0.1",
+    "@types/node": "^14.14.28",
+    "@types/webpack": "^4.41.26",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^7.0.0",
-    "css-loader": "^5.0.1",
-    "eslint": "^7.15.0",
+    "css-loader": "^5.0.2",
+    "eslint": "^7.20.0",
     "eslint-config-standard": "^16.0.2",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-promise": "^4.3.1",
     "prettier": "^2.2.1",
     "style-loader": "^2.0.0",
-    "webpack": "^5.16.0",
-    "webpack-cli": "^4.4.0"
+    "ts-loader": "^8.0.17",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.5",
+    "webpack": "^5.22.0",
+    "webpack-cli": "^4.5.0"
   },
   "dependencies": {
     "codemirror": "^5.59.2",
